Divide in hpp and cpp
