+++ CMakeLists.txt
@@ -33,12 +33,17 @@ fftw::fftw3
 
 # We put here the name of our class, and the libraries we want to link with
 
+daq_add_library(
+  algs/*.cpp
+  LINK_LIBRARIES ${DQM_DEPENDENCIES}
+)
+
 daq_add_plugin(DQMProcessor duneDAQModule LINK_LIBRARIES ${DQM_DEPENDENCIES})
 target_include_directories(dqm_DQMProcessor_duneDAQModule PUBLIC $ENV{FFTW_INC})
 
 # Unit tests
-daq_add_unit_test(Fourier_test LINK_LIBRARIES ${DQM_DEPENDENCIES})
-daq_add_unit_test(RMS_test LINK_LIBRARIES ${DQM_DEPENDENCIES})
-daq_add_unit_test(STD_test LINK_LIBRARIES ${DQM_DEPENDENCIES})
+daq_add_unit_test(Fourier_test LINK_LIBRARIES ${DQM_DEPENDENCIES} dqm)
+daq_add_unit_test(RMS_test LINK_LIBRARIES ${DQM_DEPENDENCIES} dqm)
+daq_add_unit_test(STD_test LINK_LIBRARIES ${DQM_DEPENDENCIES} dqm)
 
 daq_install()
