+++ include/dqm/algs/STD.hpp
@@ -5,8 +5,8 @@
  * Licensing/copyright details are in the COPYING file that you should have
  * received with this code.
  */
-#ifndef DQM_INCLUDE_DQM_STDEV_HPP_
-#define DQM_INCLUDE_DQM_STDEV_HPP_
+#ifndef DQM_INCLUDE_DQM_ALGS_STD_HPP_
+#define DQM_INCLUDE_DQM_ALGS_STD_HPP_
 
 #include <cmath>
 
@@ -36,37 +36,7 @@ public:
   double std() const;
 };
 
-
-void
-STD::fill(double const x)
-{
-  m_nentries++;
-  m_sum += x;
-  m_sum_sq += x * x;
-}
-
-void
-STD::clean()
-{
-  m_sum = 0;
-  m_sum_sq = 0;
-  m_nentries = 0;
-}
-
-double
-STD::std() const
-{
-  if (m_std_set) {
-    return m_std;
-  }
-  if (m_nentries <= 1) {
-    return -1;
-  }
-  auto mean = m_sum / m_nentries;
-  return sqrt((m_sum_sq + m_nentries * mean * mean - 2 * m_sum * mean) / (m_nentries - 1));
-}
-
 } // namespace dunedaq
 } // namespace dqm
 
-#endif // DQM_INCLUDE_DQM_STDEV_HPP_
+#endif // DQM_INCLUDE_DQM_ALGS_STD_HPP_
